2008-03-08  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com>
	Type clarification: size_t for length or size, uint64_t for file
	length, off_t for file offset.
	* src/DefaultBtProgressInfoFile.cc
	* src/DefaultBtContext.{h, cc}
	* src/DefaultBtInteractive.cc
	* src/DownloadCommand.{h, cc}
	* src/IteratableChunkChecksumValidator.{h, cc}
	* src/SegmentMan.{h, cc}
	* src/SingleFileDownloadContext.{h, cc}
	* src/IteratableChecksumValidator.{h, cc}
	* src/BtContext.h
	* src/IteratableValidator.h
	* src/DownloadContext.h
	* test/SingleFileDownloadContextTest.cc
	* test/DefaultBtContextTest.cc
	* test/MockBtContext.h
	* test/Metalink2RequestGroupTest.cc

+++ src/BtContext.h
@@ -54,7 +54,7 @@ public:
 
   virtual const unsigned char* getInfoHash() const = 0;
 
-  virtual int32_t getInfoHashLength() const = 0;
+  virtual size_t getInfoHashLength() const = 0;
 
   virtual std::string getInfoHashAsString() const = 0;
 
@@ -73,8 +73,8 @@ public:
     return _private;
   }
 
-  virtual std::deque<int32_t>
-  computeFastSet(const std::string& ipaddr, int32_t fastSetSize) = 0;
+  virtual std::deque<size_t>
+  computeFastSet(const std::string& ipaddr, size_t fastSetSize) = 0;
   
   virtual RequestGroup* getOwnerRequestGroup() = 0;
