+++ test/Metalink2RequestGroupTest.cc
@@ -44,7 +44,7 @@ void Metalink2RequestGroupTest::testGenerate()
     CPPUNIT_ASSERT_EQUAL(std::string("http://httphost/aria2-0.5.2.tar.bz2"), uris[1]);
     SharedHandle<SingleFileDownloadContext> dctx = rg->getDownloadContext();
     CPPUNIT_ASSERT(!dctx.isNull());
-    CPPUNIT_ASSERT_EQUAL((int64_t)0, dctx->getTotalLength());
+    CPPUNIT_ASSERT_EQUAL(0ULL, dctx->getTotalLength());
 #ifdef ENABLE_MESSAGE_DIGEST
     CPPUNIT_ASSERT_EQUAL(std::string("sha1"), dctx->getChecksumHashAlgo());
     CPPUNIT_ASSERT_EQUAL(std::string("a96cf3f0266b91d87d5124cf94326422800b627d"),
@@ -61,7 +61,7 @@ void Metalink2RequestGroupTest::testGenerate()
 #ifdef ENABLE_MESSAGE_DIGEST
     CPPUNIT_ASSERT_EQUAL(std::string("sha1"), dctx->getPieceHashAlgo());
     CPPUNIT_ASSERT_EQUAL((size_t)2, dctx->getPieceHashes().size());
-    CPPUNIT_ASSERT_EQUAL((int32_t)262144, dctx->getPieceLength());
+    CPPUNIT_ASSERT_EQUAL((size_t)262144, dctx->getPieceLength());
     CPPUNIT_ASSERT_EQUAL(std::string(""), dctx->getChecksumHashAlgo());
     CPPUNIT_ASSERT_EQUAL(std::string(""), dctx->getChecksum());
 #endif // ENABLE_MESSAGE_DIGEST
